修复Swiper 超长显示导航点快速滑动时，导航点区域闪现蒙层问题

+++ frameworks/core/components_ng/pattern/swiper_indicator/dot_indicator/overlength_dot_indicator_modifier.h
@@ -203,7 +203,7 @@ private:
     void UpdateSelectedCenterXOnDrag(const LinearVector<float>& itemHalfSizes);
     void UpdateUnselectedCenterXOnDrag();
     int32_t CalcTargetIndexOnDrag() const;
-    std::pair<float, float> CalcLongPointEndCenterXWithBlack(int32_t index, const LinearVector<float>& itemHalfSizes);
+    std::pair<float, float> CalcLongPointEndCenterXWithBlack(size_t index, const LinearVector<float>& itemHalfSizes);
     float GetMoveRateOnAllMove() const;
     int32_t GetBlackPointsAnimationDuration() const;
     void AdjustTargetStatus(int32_t targetPageIndex);
